+++ src/Fields/index.ts
@@ -1,4 +1,8 @@
+export * from './AutocompleteField';
 export * from './CheckboxField';
+export * from './DatePickerField';
+export * from './DateRangePickerField';
+export * from './CheckboxGroupField';
 export * from './FieldProvider';
 export * from './FieldWrapper';
 export * from './formItemFieldProps';
@@ -6,5 +10,6 @@ export * from './InputField';
 export * from './InputNumberField';
 export * from './RadioGroupField';
 export * from './SelectField';
+export * from './SwitchField';
 export * from './TextAreaField';
 export * from './types';