+++ src/Space/index.stories.tsx
@@ -30,33 +30,33 @@ export const Default: Story<SpaceProps> = function Default(args) {
 
 export const Block: Story<SpaceProps> = function Block(args) {
   return (
-    <Space direction="vertical" block>
+    <Space vertical block>
       <Card title="Vertical space without block">
-        <Space {...args} direction="vertical">
+        <Space {...args} vertical>
           <Button block>Button1</Button>
           <Button block>Button2</Button>
         </Space>
       </Card>
       <Card title="Vertical space with block">
-        <Space {...args} direction="vertical" block>
+        <Space {...args} vertical block>
           <Button block>Button1</Button>
           <Button block>Button2</Button>
         </Space>
       </Card>
       <Card title="Vertical space with block and align='center'">
-        <Space {...args} direction="vertical" block align="center">
+        <Space {...args} vertical block align="center">
           <Button block>Button1</Button>
           <Button block>Button2</Button>
         </Space>
       </Card>
       <Card title="Horizontal space without block">
-        <Space {...args} direction="horizontal">
+        <Space {...args}>
           <Button block>Button1</Button>
           <Button block>Button2</Button>
         </Space>
       </Card>
       <Card title="Horizontal space with block (no difference because not horizontally centered)">
-        <Space {...args} direction="horizontal" block>
+        <Space {...args} vertical={false} block>
           <Button block>Button1</Button>
           <Button block>Button2</Button>
         </Space>