+++ src/Space/index.tsx
@@ -4,17 +4,22 @@ import styled, { css } from 'styled-components';
 
 export { SpaceSize } from 'antd/lib/space';
 
-export type SpaceProps = AntdSpaceProps & {
+export type SpaceProps = Omit<AntdSpaceProps, 'direction'> & {
   block?: boolean;
+  vertical?: boolean;
 };
 
-const TRANSIENT_PROPS: Array<string> = ['block'] satisfies Array<
+const TRANSIENT_PROPS: Array<string> = ['block', 'vertical'] satisfies Array<
   keyof SpaceProps
 >;
 
-export const Space: React.FC<SpaceProps> = styled(AntdSpace).withConfig({
-  shouldForwardProp: (prop) => !TRANSIENT_PROPS.includes(prop),
-})`
+export const Space: React.FC<SpaceProps> = styled(AntdSpace)
+  .withConfig({
+    shouldForwardProp: (prop) => !TRANSIENT_PROPS.includes(prop),
+  })
+  .attrs<SpaceProps>((props) => ({
+    direction: props.vertical ? 'vertical' : 'horizontal',
+  }))`
   ${(props: SpaceProps) =>
     props.block &&
     css`