+++ src/index.ts
@@ -6,8 +6,9 @@ export * from './theme';
 export * from './Provider';
 
 // Individual components
-export * from './Anchor';
 export * from './Alert';
+export * from './Anchor';
+export * from './Autocomplete';
 export * from './Avatar';
 export * from './BackTop';
 export * from './Badge';
@@ -20,6 +21,7 @@ export * from './Comment';
 export * from './DatePicker';
 export * from './Descriptions';
 export * from './Divider';
+export * from './Drawer';
 export * from './Dropdown';
 export * from './Empty';
 export * from './Fields'; // Might extract to separate package later due to dependency on react-hook-form